requirements.txt
spotipy
wifi
pillow
qrcode
flask